Refactor for authorization in Auth app
