--- auth/src/main/java/ates/homework/auth/entity/BookEntity.java
@@ -1,20 +1,0 @@
-package ates.homework.auth.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "books")
-@NoArgsConstructor
-@Getter
-@Setter
-public class BookEntity {
-    @Id
-    private long id;
-    private String title;
-    private String author;
-}
