--- auth/src/main/java/ates/homework/auth/repository/BookRepository.java
@@ -1,9 +1,0 @@
-package ates.homework.auth.repository;
-
-import ates.homework.auth.entity.BookEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BookRepository extends JpaRepository<BookEntity, Long> {
-}
